@@ -6,12 +6,14 @@
     "@testing-library/jest-dom": "^5.11.6",
     "@testing-library/react": "^11.2.2",
     "@testing-library/user-event": "^12.5.0",
+    "axios": "^0.27.2",
     "bootstrap": "^4.5.3",
     "emailjs-com": "^2.6.4",
     "html-react-parser": "^1.0.0",
-    "react": "^17.0.1",
+    "react": "^18.2.0",
+    "react-alert": "^7.0.3",
     "react-bootstrap": "^1.4.0",
-    "react-dom": "^17.0.1",
+    "react-dom": "^18.2.0",
     "react-hook-form": "^6.13.1",
     "react-if-elseif-else-render": "^1.0.2",
     "react-modal": "^3.12.1",
