@@ -1,10 +1,19 @@
 import React from 'react';
-import ReactDOM from 'react-dom';
+import ReactDOM from 'react-dom/client';
 import App from './components/App';
+import { positions, transitions, Provider as AlertProvider } from 'react-alert';
+import AlertTemplate from 'react-alert-template-basic';
 
+const options = {
+  duration: 5000,
+  position: positions.MIDDLE,
+  transition: transitions.SCALE
+}
 
-ReactDOM.render(
-  
-  <App />, document.getElementById('root')
+const root = ReactDOM.createRoot(document.getElementById('root'));
+root.render(
+  <AlertProvider template={AlertTemplate} {...options}>
+    <App />
+  </AlertProvider>
 );
 
