@@ -1,21 +1,19 @@
 import React from 'react';
 
-
-
-function Banner() {
+const Banner = () => {
     return (
 
 <div id="home">
 {/* <!-- .container --> */}
-<div class="container-fluid" id="banner-fluid">
+<div className="container-fluid" id="banner-fluid">
 
 {/* <!-- START THE FEATURETTES -->
   <!-- Wrap the rest of the page in another container to center all the content. --> */}
-  <div class="row featurette row-content">
-    <div class="col-lg-5">
+  <div className="row featurette row-content">
+    <div className="col-lg-5">
       <br />
-      <h2 class="featurette-heading">COMPANY OVERVIEW</h2>
-      <p class="lead">We were founded in 2020 and our headquarter is in Toronto, Ontario. 
+      <h2 className="featurette-heading">COMPANY OVERVIEW</h2>
+      <p className="lead">We were founded in 2020 and our headquarter is in Toronto, Ontario. 
       We take pride to call ourselves one of the pioneer companies offering IT Consulting 
       Services worldwide. You name the technology and we have it in store for you. We are a 
       team of highly enthusiastic designers, developers, technical analysts and digital 
@@ -23,7 +21,7 @@ function Banner() {
       </p>
       <br />
     </div>
-    <div class="col-md-7">
+    <div className="col-md-7">
     </div>
   </div>
 </div>
