@@ -1,27 +1,33 @@
 import React from 'react';
-import {Navbar, Nav, Container} from 'react-bootstrap';
-import { Link, animateScroll as scroll } from "react-scroll";
+import { Navbar, Nav } from 'react-bootstrap';
+import { Link } from "react-scroll";
 
-function Header() {
+const Header = () => {
   return (
     <>
-      <Navbar className="container-fluid sticky-top" collapseOnSelect expand="lg" bg="dark" variant="dark" fixed-top>
-        {/* <Container> */}
+      <Navbar className="container-fluid sticky-top" 
+              collapseOnSelect 
+              expand="lg" 
+              bg="dark" 
+              variant="dark" 
+              fixed-top>
         <Navbar.Brand className="mr-auto" >
           <Link 
-              to ="home"
+              to ="/home"
               activeClass="active"
               className="nav-link"
               spy={true}
               smooth={true}
               duration={500}
-              offset = {-100}
-              >
+              offset = {-100}>
 
-          <img src="./img/Logo.png" width="160px" className="d-inline-block align-top" 
-          alt="logo of the TechSup"/>
+              <img src="./img/Logo.png" 
+                  width="160px" 
+                  className="d-inline-block align-top" 
+                  alt="logo of the TechSup"/>
           </Link>
         </Navbar.Brand>
+
         <Navbar.Toggle aria-controls="responsive-navbar-nav" />
         <Navbar.Collapse id="responsive-navbar-nav" >
           <Nav className="ml-auto" fluid>
@@ -71,7 +77,6 @@ function Header() {
           </Link>
           </Nav>
         </Navbar.Collapse>
-        {/* </Container> */}
       </Navbar>
     </>
   );
