@@ -1,17 +1,17 @@
 import React from 'react';
 
-function Profile() {
+const Profile = () => {
     return (
 // <!-- .container -->
-<div class="container-fluid" id="portfolio"> 
+<div className="container-fluid" id="portfolio"> 
 
 {/* 
-  <!-- <hr class="featurette-divider"> -->
+  <!-- <hr className="featurette-divider"> -->
   <!-- Portfolio --> */}
-  <div class="center row row-content">
+  <div className="center row row-content">
 
-    <div class="col-lg-6">
-     <div class="profile_img">
+    <div className="col-lg-6">
+     <div className="profile_img">
         <img src="./img/miranda.jfif" width="140px" height="140px" alt="Miranda"/>
       </div>
       <h3 id="ourName">Miranda Qi</h3>
@@ -22,12 +22,12 @@ function Profile() {
         capital markets, insurance, manufacturing, and energy. Miranda has a BSc. in Psychology 
         and a Diploma in Computer Programming.
       </p>
-      <p><a class="btn btn-secondary" href="https://www.linkedin.com/in/miranda-xinying-qi-33821795/" role="button">View details &raquo;</a></p>
+      <p><a className="btn btn-secondary" href="https://www.linkedin.com/in/miranda-xinying-qi-33821795/" role="button">View details &raquo;</a></p>
     </div>
     
     {/* <!-- /.col-lg-4 --> */}
-    <div class="col-lg-6">
-      <div class="profile_img">
+    <div className="col-lg-6">
+      <div className="profile_img">
         <img src="./img/mehran.jfif" width="140px" height="140px" alt="Mehran"/>
       </div>
       <h3 id="ourName">Mehran Alipour</h3>
@@ -39,7 +39,7 @@ function Profile() {
         the ultimate website for you.
       </p>
       <br/>
-      <p><a class="btn btn-secondary" href="https://www.linkedin.com/in/mehran-alipour-3a5b1a14a/" role="button">View details &raquo;</a></p>
+      <p><a className="btn btn-secondary" href="https://www.linkedin.com/in/mehran-alipour-3a5b1a14a/" role="button">View details &raquo;</a></p>
     </div>
     {/* <!-- /.col-lg-4 --> */}
     
@@ -48,9 +48,6 @@ function Profile() {
 
 </div>
 // <!-- .container -->
-// </div>
-
-        );
-    }
+)}
    
-   export default Profile;
+export default Profile;
